minor formatting

+++ src/notification.c
@@ -418,7 +418,7 @@ struct notification *notification_create(void)
         n->ellipsize = PANGO_ELLIPSIZE_MIDDLE;
         n->alignment = PANGO_ALIGN_LEFT;
         n->hide_text = false;
-        n->icon_position = (int)settings.icon_position;
+        n->icon_position = settings.icon_position;
         n->icon_size = 32;
 
         n->script_run = false;
